macOS: make `makeWindowFullyTransparent` non-async
`makeWindowFullyTransparent` was initially unnecessarily async as to not
introduce a breaking change if it ever needed to be async in the future,
when a bug associated with it was fixed.
However, it is unlikely that fixing that bug would result in the method
needing to become async, therefore, this commit removes this method's
async property.